EMVCard.cpp: ArrayToHexWideString: fix of the comparison of different types
../Common/EMVCard.cpp: In function 'std::wstring VeraCrypt::ArrayToHexWideString(con
st unsigned char*, size_t)':
../Common/EMVCard.cpp:32:43: warning: comparison of integer expressions of different
 signedness: 'int' and 'size_t' {aka 'long unsigned int'} [-Wsign-compare]
   32 |                         for (int i = 0; i < cbData; i++)
      |                                         ~~^~~~~~~~

+++ src/Common/EMVCard.cpp
@@ -29,7 +29,7 @@ namespace VeraCrypt
 		wstring result;
 		if (pbData)
 		{
-			for (int i = 0; i < cbData; i++)
+			for (size_t i = 0; i < cbData; i++)
 			{
 				result += hexChar[pbData[i] >> 4];
 				result += hexChar[pbData[i] & 0x0F];
